requirements.txt
Pillow==11.2.1
pyautogui==0.9.54
pynput==1.8.1
schedule==1.2.2
